pyproject.toml
[project]
name = "screensaver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Zaher Ghaibeh", email = "zaher@zaher.dev" }
]
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.19",
    "pillow>=12.0.0",
    "platformdirs>=4.5.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.21.0",
]

[project.scripts]
screensaver = "screensaver:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 90

[tool.ruff.lint]
select = ["E", "F", "I", "Q", "W"]
